requirements.txt
atomicwrites==1.4.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.11
colorama==0.4.4
coverage==5.2
defusedxml==0.7.1
Django==2.2.9
django-cors-headers==3.11.0
django-debug-toolbar==3.2.4
django-extensions==3.1.5
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
factory-boy==2.12.0
Faker==11.3.0
graphviz==0.19.1
httpie==3.0.2
idna==3.3
iniconfig==1.1.1
MarkupSafe==2.0.1
multidict==6.0.2
ngrok==0.0.1
packaging==21.3
Pillow==9.0.0
pluggy==1.0.0
py==1.11.0
pydot==1.4.2
Pygments==2.11.2
PyJWT==2.3.0
pyparsing==3.0.6
PySocks==1.7.1
pytest==6.2.5
pytest-django==3.8.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
requests-toolbelt==0.9.1
six==1.16.0
sorl-thumbnail==12.7.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.8
